Add user existence check in RequestController
Implemented a check to verify if a user exists based on the UserId from the requestDto. If the user is not found, a BadRequest response is returned. The previous user existence check has been commented out. Updated the SignalR notification to use the correct username from the loaded user object.